Move supplyLayoutDefaults to own file
